data-collector/requirements.txt
requests
beautifulsoup4
pandas
openpyxl
yfinance
pandas_datareader
python-dotenv
psycopg2-binary
